+++ uninstall/uninstall.sh
@@ -2,4 +2,5 @@
 set -e
 
 rm -f /etc/sudoers.d/loxberryhostbackup
+rm -f /etc/cron.d/loxberryhostbackup
 exit 0